Don't decr profiling count without incr
The failure cleanup code in dvmMethodTraceStart() was assuming
that profiling had already been started, which was true until my
previous change.  The code in the interpreter isn't too picky, but
updateActiveProfilers() is trying to keep an accurate acount, and
gets upset.  If you pass in a bad filename, we disable something that
was never enabled, and the VM gets confused and aborts out of spite.

Related to bug 5564440

Change-Id: I17fd8bf4654c29f49efb10142d222a91c190b6ed