Disable ADDR_NO_RANDOMIZE for 3.4 or greater kernels
Early kernels had memory fragmentation bugs on ARM when
ASLR was enabled. The ADDR_NO_RANDOMIZE flag was added
to work around this bug until the appropriate kernel fix
could be made.

Linux kernel 3.4 seems to have this fix, making this workaround
unnecessary.  Remove the workaround if we detect we're running
an up-to-date kernel.

(I believe the kernel patch is also in 3.3, but haven't explicitly
tested it).

Bug: 5817320
Change-Id: Ifcbc68fdfb2ada0f55358e7a7af9e155fe3bec4f