Fix broken card table asserts.
We had an off by one error due to getting the card for the heap
source limit. This is not a valid card when the heap is at
maximum size, but doesn't need to be since we do less than
comparison on it. Fixed another error where we didn't take into
account the biased begin in an assert.

Did a bit of refactoring by removing useless if statement.

Fixes 061-out-of-memory.

Bug: 8282011
Change-Id: I66e3026e7b98b738ecfbced101846ec9f184ab70