Revert "Remove some of the x86 libdvm.so text relocations."
This reverts commit 469e580f6a18c00a3c356ae3898aa8d143736016.

Although dalvik/tests/run-all-tests passed on an emulator after this change, starting a host VM crashes in dvmAsmInstructionStartCode, which is highly suspicious.

Since this change didn't actually solve my problem (I'd need to remove all text relocations to make the linker happy with the x86 libdvm.so, so it would stop warning, so the tests would stop failing because the output doesn't match the golden output), I'm just going to revert and leave it for now.

Change-Id: Ic35acdb0ae9a5b321ba728afceff1c1d9b0114fc