Fix mainDexClasses with obfuscation.
When the input jar is obfuscated, multidex package can be renamed so
the multidex classes were not always correctly included in the
main dex. Tracing from Application.attachBaseContext() is not
subject to this problem since the method is defined in android API
and this will keep the multidex clases in the main dex since
attachBaseContext must call MultiDex.install().

Change-Id: I62efd2292700a04c2bc3fd7434d20b18d3d74343