Do not generate empty dex in multidex.
This was possible when the first class considered for adding in a dex was too
big to be sure it will fit. Now DX will try to add the class in the current dex
and crash if the class causes one index to overflow.

(cherry picked from commit f15dafe674d2270cafaf4fb7993963ec474d3761)

Change-Id: I323f8111c2be6bace908d0539ff324c4bd033c05