Fixes command line too long error on Windows
On Windows machines the command line can only accept
commands that are 8192 characters or less.  This can
cause a problem when a project has many dependencies
and many libraries are being passed to dex at once.

This fixes it by adding a new option to dex called
--input-list that can read inputs from a file.
This way we only have to pass a single file name
instead of a ton of jar files. This keeps the command
short so Windows will not complain about it.

Bug: 17438982

Signed-off-by: John Fazekas <jfazekas@fd-software.com>

(cherry picked from commit 914f2263ca57b21dc4ef1fb218e345a16a5a5db5)

Change-Id: I6edd1e7c1d1fddc8d6bf4c4963295d4f4e64032e