Sort LocalVariableList
To retrieve the local variable info that corresponds to a register from the
pc, the LocalVariableList must be sorted by pc.
This is because we match variables using the end pc and not the start pc
anymore (which is unreliable, see 455443f47d9e217b99edcc33202f889e2c637b7b),
so for a given index we need to try the LV info with the lower pc first.

Test: manual
Bug: 32432143
Change-Id: I6289f9a2103b99d8362cf6870e45d3fb342d32ca