Fix bad instruction massaging
- The same register can only be used between the result register and
the first operand register of an instruction but must not be used
between the first operand and the second operand of an instruction.
We must reset the value after handling the first register, whether it
needs to be replaced or not.
This is the same CL done for dx as Jack's backend in
a9d086fd8b99fb97173f45b357547b14a5a6a561

Bug: 34624820

Test: manually
Change-Id: Ia5bfaa78f90ff1086fbe0581cb04186332d33ae7