Refinements relating to invoke-custom
Move assignment of call site references in dx to avoid call site
duplication. Previously we were allocating call site references in the
BytecodeArray when visiting instructions, but there are two visitors
that are applied to instructions :- the BasicBlocker and the
Simulator. Allocation now happens in the Simulator visitor.

Update dexdump to print call site offset of each listed callsite to
enable calculating the number of call site ids and call site instances.

Update 135-invoke-custom test to sanity check the number of call sites
and call site ids.

Do not allow dx/tests/run-test to report success if it fails when run
with --update.

Bug: 36641082
Test: dalvik/dx/tests/run-all-tests
Test: dalvik/dx/tests/run-test 135
Change-Id: I8bb1f6c99c97c7a9fa785503df4f66c9e9b80672