Write zeros to byte array output
Because we reuse the output array
between seprate dx invocations, we
must write zeros to array. Otherwise, we might
end up picking up the output from
previous invocation.

Dx had an assumption that a new output
array is always allocated, and it was
a performance optimization not to write
zeros out. That does not apply any more.

BUG=65363025
Test: ByteArrayAnnotatedOutputTest

Change-Id: I984076e8078f9ed977f10cacb9b1b135b1d6358e