libdex: Enable -Wimplicit-fallthrough
This adds fallthrough annotations to this code, to make it compile
cleanly when given -Wimplicit-fallthrough. The intent is to enable this
warning for the Android platform after these uses are cleaned up.

This code runs on the host, rather than the device, so we can't rely on
the FALLTHROUGH_INTENDED macro provided by android-base here. Since
said macro is pretty trivial, duplicating it doesn't seem all that bad.

Bug: 112564944
Test: Builds
Change-Id: I9f49f15d5dd178a2d78a8092d5863a42369968c0